INITIALIZE REPO
This sandbox aims to showcase Phoenix, Ecto, and PostgreSQL all using ULIDs style UUIDS in harmony with each other.

---  How it works  ---

The `pg-ulid` extension provides a unique field type of `ulid` and is specified as-such when the database provides the
data to the application.

The `pg-ulid` extension stores the ULID identifier in the database as a binary data (BYTEA) field.

The ULID is cast into a string when being displayed or queried from the database.

The Ecto.ULID module handles the ulid field data as a String, as that's the format provided by the database.

The Ecto.ULID module auto-generates a valid ULID upon persisting new %Structs{}. But does NOT do so when using Repo
functions that bypass callbacks such as `Repo.insert_all/3`.

In situations where `Repo.insert_all/3` needs to be used, postgres can generate it's own ULIDs natively with `pg-ulid`.